src/gpio/requirements.txt
RPi.GPIO==0.7.1
spidev==3.6
mfrc522==0.0.7
smbus2==0.4.1
RPLCD==1.4.0
flask==3.0.2
requests==2.31.0